(PC-19515)[PRO] refactor: rename stepper to breadcrumb
