+++ pro/src/components/OfferIndividualBreadcrumb/index.ts
@@ -1,0 +1,2 @@
+export { default as OfferIndividualBreadcrumb } from './OfferIndividualBreadcrumb'
+export { OFFER_WIZARD_STEP_IDS } from './constants'
