--- pro/src/components/OfferIndividualStepper/index.ts
@@ -1,2 +1,0 @@
-export { default as OfferIndividualStepper } from './OfferIndividualStepper'
-export { OFFER_WIZARD_STEP_IDS } from './constants'
