+++ pro/src/core/Offers/utils/getOfferIndividualUrl/__specs__/getOfferIndividualUrl.spec.ts
@@ -1,4 +1,4 @@
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import { OFFER_WIZARD_MODE } from 'core/Offers'
 
 import getOfferIndividualUrl from '../getOfferIndividualUrl'
