+++ pro/src/core/Offers/utils/getOfferIndividualUrl/getOfferIndividualUrl.ts
@@ -1,6 +1,6 @@
 import { generatePath } from 'react-router'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import { OFFER_WIZARD_MODE } from 'core/Offers'
 
 import {
