+++ pro/src/core/Offers/utils/getOfferIndividualUrl/getOfferIndividualUrlV2/getOfferIndividualUrlV2.ts
@@ -1,6 +1,6 @@
 import { generatePath } from 'react-router'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import { OFFER_WIZARD_MODE } from 'core/Offers'
 
 interface IGetOfferIndividualPathArgs {
