+++ pro/src/pages/Offers/Offers/OfferItem/Cells/EditStocksCell/EditStocksCell.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import {
   Events,
   OFFER_FORM_NAVIGATION_IN,
