+++ pro/src/screens/OfferIndividual/StocksEvent/__specs__/StocksEvent.RouteLeavingGuard.spec.tsx
@@ -12,7 +12,7 @@ import {
   StockResponseModel,
 } from 'apiClient/v1'
 import Notification from 'components/Notification/Notification'
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import {
   IOfferIndividualContext,
   OfferIndividualContext,
