+++ pro/src/screens/OfferIndividual/StocksThing/__specs__/StocksThing.RouteLeavingGuard.spec.tsx
@@ -11,7 +11,7 @@ import {
   GetIndividualOfferResponseModel,
   StockResponseModel,
 } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import {
   IOfferIndividualContext,
   OfferIndividualContext,
