+++ pro/src/screens/OfferIndividual/Summary/StockSection/__specs__/StockSection.spec.tsx
@@ -7,7 +7,7 @@ import { Provider } from 'react-redux'
 import { generatePath, MemoryRouter, Route } from 'react-router'
 
 import { OfferStatus } from 'apiClient/v1'
-import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualStepper'
+import { OFFER_WIZARD_STEP_IDS } from 'components/OfferIndividualBreadcrumb'
 import { OFFER_WIZARD_MODE } from 'core/Offers'
 import { getOfferIndividualPath } from 'core/Offers/utils/getOfferIndividualUrl'
 import * as useAnalytics from 'hooks/useAnalytics'
